GUI proto (only Record tab); Recorder facility implemented; Fixed wrong packing process - packed chunk was not append to output buffer;

+++ src/modes/modes-data.h
@@ -13,14 +13,14 @@ public:
     ModesData(): m_type{MessageType::None}, m_timestamp{0} {
     }
     void loadMessage(MessageType, const QByteArray &);
-    inline unsigned timestamp() const {
+    inline quint64 timestamp() const {
         return m_timestamp;
     }
     QByteArray serialize() const;
 private:
     void loadMessageBeast(const QByteArray &);
     MessageType m_type;
-    unsigned m_timestamp;
+    quint64 m_timestamp;
     QByteArray m_data; // complete AVR frame
 };
